Extract RV32I/RV64I tests into dedicated files
Move base integer instruction tests from TestLifting32.cpp and
TestLifting64.cpp into RV32I.cpp and RV64I.cpp respectively, matching
the pattern used by other extensions (M, A, C, F, D, Zicsr). The
TestLifting files now contain only the main() entry point.
